fix: remove `ManagedPoolController` from the repo
