+++ packages/balancer/agents/autonomous_fund/tests/helpers/constants.py
@@ -48,8 +48,7 @@ COMPATIBILITY_FALLBACK_HANDLER = "0x559E01ac5e8fe78963998D632e510bEF3e306A78"
 CREATE_CALL = "0x1967D06b1fabA91eAadb1be33b277447ea24fa0e"
 MULTISEND = "0x336e71DaB0302774b1e4c53202bF3f2D1aD1a8e6"
 MULTISEND_CALL_ONLY = "0x3635D6aE8610Ea00b6AD8342b819fD21c7Db77Ed"
-MANAGED_POOL_CONTROLLER = "0x28BF8d29cFA99aE9C3D876210453272f30e4D131"
-MANAGED_POOL = "0x1Aec61C3CA6A41B8114e00C29FDC36EF095Fe1A4"
+MANAGED_POOL = "0xb5f3FC2579b134D836271AC872de2DA83Fe6e6a1"
 MANAGED_POOL_TOKENS = [
     "0x3A78BF1783a0187c1C8000e41C2a008897D0a35f",
     "0xC97b465daC9f52A26F2A234c658a57f5B3f15D19",
