+++ packages/balancer/contracts/managed_pool/contract.yaml
@@ -8,7 +8,7 @@ aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeifuqjdhsrydlwgdmuzlwvxsidbvxqrz3ulec4usczuquaj3g6jn4u
   build/IManagedPool.json: bafybeidmc2exztkruyzk3ohwzfclsemhcnfalt6kezwmgnmtj2efcprk4q
-  contract.py: bafybeicdydetjn7bi4reyska4algtwepfo5udkhi643unn7qe6ayg4ppfq
+  contract.py: bafybeidr7w26utcsykiglkgl3ago4nar4re3abacayglicfxzalr5iaujq
 fingerprint_ignore_patterns: []
 contracts: []
 class_name: ManagedPoolContract
