--- packages/balancer/contracts/managed_pool_controller/contract.yaml
@@ -1,21 +1,0 @@
-name: managed_pool_controller
-author: balancer
-version: 0.1.0
-type: contract
-description: The contract package for the balancer ManagedPoolController.
-license: Apache-2.0
-aea_version: '>=1.0.0, <2.0.0'
-fingerprint:
-  __init__.py: bafybeib6wwdtqj6k65hwjampb37d4posy7lxik3ry7m7e63d7hxpecn5au
-  build/ManagedPoolController.json: bafybeiedp7s3hmdex7z35cqavk4whpglztxhquilopy3u63stpwq2qmomu
-  contract.py: bafybeih5vtthccrbbxap2s7xqqs2kvmniixfb6ijdaywe5ekfkxgoa5k5u
-fingerprint_ignore_patterns: []
-contracts: []
-class_name: ManagedPoolControllerContract
-contract_interface_paths:
-  ethereum: build/ManagedPoolController.json
-dependencies:
-  open-aea-ledger-ethereum:
-    version: ==1.26.0
-  web3:
-    version: ==5.25.0
