+++ packages/balancer/skills/autonomous_fund_abci/skill.yaml
@@ -24,7 +24,7 @@ contracts: []
 protocols: []
 skills:
 - balancer/fear_and_greed_oracle_abci:0.1.0:bafybeigwp2sg7llmhzddjqf2ixbpt6g3lsvlcvog2ruf2fk523d2nnjp4u
-- balancer/pool_manager_abci:0.1.0:bafybeiefrle4pcjeb34ferof5iq6wmvxnhiljoru3rfzeqotfqyjbc5yk4
+- balancer/pool_manager_abci:0.1.0:bafybeieh64tn37w6y5hhhrrqdrvrqrxjs4unafdgo3m4k7coh5aotpowsy
 - valory/abstract_round_abci:0.1.0:bafybeih7sibc3evofk3fnmfijavyp6hhkku7fg4l336qysqwlnhvjhsvlm
 - valory/registration_abci:0.1.0:bafybeibvuyylbaxqits6eskh5i3wrtyglr4l7gmahvcaqevphv7haj5mbi
 - valory/reset_pause_abci:0.1.0:bafybeibq5tl6xf4g22ojzzeglylbanu5nxfsf6s6g7uf3yubogpv3acc4q
@@ -100,8 +100,7 @@ models:
       history_check_timeout: 1205
       keeper_allowed_retries: 3
       keeper_timeout: 30.0
-      managed_pool_address: '0x28BF8d29cFA99aE9C3D876210453272f30e4D131'
-      managed_pool_controller_address: '0x28BF8d29cFA99aE9C3D876210453272f30e4D131'
+      managed_pool_address: '0xb5f3FC2579b134D836271AC872de2DA83Fe6e6a1'
       max_healthcheck: 120
       max_index_change: 0.00040509259
       max_index_value: 100
