+++ packages/balancer/skills/pool_manager_abci/skill.yaml
@@ -7,15 +7,15 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeib7tq4xbys4lmfpz4kp7qs6vcw5lifiq6zb2a4strssd256g7rici
-  behaviours.py: bafybeihjgyk3tufxzbzlj62h43fxkh7eg563dv52zhx4sjxmq7wujc6fki
+  behaviours.py: bafybeicevqpesafwlk4qpdknilgfrhynxdi3mnkmdxbj7g6np2fdaowoxi
   dialogues.py: bafybeic2grmbb2eepv745wuv6gbyx6xwahipduw6bzqbt5vcj3h6cpln7q
   fsm_specification.yaml: bafybeify5zvsgb6acwgcekmltgb4rlpeovls4plcqmcuoyr7tgjlwp72pi
   handlers.py: bafybeiczztxmybnfcgggaw2dixwg6z6ql3rkicb5fstoo3iywliygpcw3u
-  models.py: bafybeiddpb3swlag6brr6c2v66o2kd7cvdri56pvufrxzmvyx2zjvcrnia
+  models.py: bafybeieuopnk4uektmr42vtnw6yydeh7ca2522kvanmb5xczwpexbo5rte
   payloads.py: bafybeihzuwv3amtvpvu2qged5ifwy746focexq4wtp75rvgbkoxp57jyrm
   rounds.py: bafybeieji2ttkd5o2rqspeygfwv6n5dyeaa5pf4cxixv3cckqfh5a2yh3e
   tests/__init__.py: bafybeiedv5h6wd6jjxzsu5d7foz3jyvgwtyotgtx2awmp2gqfysgtyziu4
-  tests/test_behaviours.py: bafybeiax6mpzqjlnsbfzya7qrasjbiqo2t7jb3gdvubo4kmd5i5knxbhxa
+  tests/test_behaviours.py: bafybeigovooer34273ot3x3srltgqxrrhg5hi6yio4ddkvkfg2xwwx3z6a
   tests/test_dialogues.py: bafybeigv2j6zxox7pk4gfpgvsdbugofglut4gmlpf3uq4hlbx64hscv7n4
   tests/test_handlers.py: bafybeiesvxlswqsmyufa52nuxbxet7oqbtbajb6xwsdmfdrf5x6wcwe47y
   tests/test_models.py: bafybeihp2476o2dteoitn27yqypucgsn7e4bdvump53d4z3ht33peucpm4
@@ -24,8 +24,7 @@ fingerprint:
 fingerprint_ignore_patterns: []
 connections: []
 contracts:
-- balancer/managed_pool:0.1.0:bafybeigwhw2pbqnjqueewtalm2dktvrma7lyfzlv75tsnzkoocavamvxh4
-- balancer/managed_pool_controller:0.1.0:bafybeihzha5pmyqff5jovqhzgdomhiajwvfbvh7swgt64p62cwfywqybme
+- balancer/managed_pool:0.1.0:bafybeiefed67esjr6q43wyjmfijrnpo2cubhwcemomzd7z4l7hzxhky6im
 - valory/gnosis_safe:0.1.0:bafybeich2izgpdvzzoodlmomvmob6yxadji6h7sp65dl4hz6vus7ra7jjm
 protocols:
 - valory/contract_api:1.0.0:bafybeiaxbrvgtbdrh4lslskuxyp4awyr4whcx3nqq5yrr6vimzsxg5dy64
@@ -97,8 +96,7 @@ models:
           version: {}
         voting_power: '10'
       keeper_timeout: 30.0
-      managed_pool_address: '0x28BF8d29cFA99aE9C3D876210453272f30e4D131'
-      managed_pool_controller_address: '0xb821BFfE924E18F8B3d92473C5279d60F0Dfc6eA'
+      managed_pool_address: '0xb5f3FC2579b134D836271AC872de2DA83Fe6e6a1'
       max_healthcheck: 120
       observation_aggregator_function: median
       observation_interval: 10
